// src/app/api/model/calendarData.ts
/**
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements. See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership. The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License. You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied. See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */

/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { LocalTime } from './localTime';
import { EnumOptionData } from './enumOptionData';


export interface CalendarData { 
    calendarInstanceId?: number;
    calendarTypeOptions?: Array<EnumOptionData>;
    centerId?: string;
    createdByUserId?: number;
    createdByUsername?: string;
    createdDate?: string;
    dateFormat?: string;
    description?: string;
    duration?: number;
    endDate?: string;
    entityId?: number;
    entityType?: EnumOptionData;
    entityTypeOptions?: Array<EnumOptionData>;
    firstReminder?: number;
    frequency?: EnumOptionData;
    frequencyNthDayTypeOptions?: Array<EnumOptionData>;
    frequencyOptions?: Array<EnumOptionData>;
    humanReadable?: string;
    id?: number;
    interval?: number;
    lastUpdatedByUserId?: number;
    lastUpdatedByUsername?: string;
    lastUpdatedDate?: string;
    locale?: string;
    location?: string;
    meetingTime?: LocalTime;
    nextTenRecurringDates?: Array<string>;
    recentEligibleMeetingDate?: string;
    recurrence?: string;
    recurringDates?: Array<string>;
    remindBy?: EnumOptionData;
    remindByOptions?: Array<EnumOptionData>;
    repeating?: boolean;
    repeatsOnDay?: EnumOptionData;
    repeatsOnDayOfMonth?: number;
    repeatsOnDayOptions?: Array<EnumOptionData>;
    repeatsOnNthDayOfMonth?: EnumOptionData;
    rowIndex?: number;
    secondReminder?: number;
    startDate?: string;
    title?: string;
    type?: EnumOptionData;
    typeId?: string;
}